requirements.txt
click>=8.1.0
requests>=2.31.0
PyPDF2>=3.0.0
pikepdf>=8.0.0
Pillow>=10.0.0
pyyaml>=6.0
questionary>=2.0.0
